Реализован функционал,который проверяет наличие фото от бригадира, кнопки для редактирования отвечают ссылкой

+++ source/tgbot/handlers/utils.py
@@ -20,11 +20,15 @@ def is_staff_in_order(func: Callable):
     def command_func(update, context, *args, **kwargs):
         chat_id = update.callback_query.message.chat.id
         message_id = update.callback_query.message.message_id
-        try:
-            call, order_id, staff_id = update.callback_query.data.split(" ")
-            Order.objects.get(pk=order_id).order_cleaners.get(staff=staff_id)
-            return func(update, context,  *args, **kwargs)
-        except:
+        call, order_id, staff_id = update.callback_query.data.split(" ")
+        if Order.objects.get(pk=order_id).order_cleaners.filter(staff=staff_id):
+            try:
+                return func(update, context, *args, **kwargs)
+            except:
+                text = "Ошибка! в коде" #Изменить текст ошибки
+                context.bot.send_message(chat_id=chat_id, text=text)
+                return
+        else:
             text = f"Вы были удалены с данного заказа!"
             context.bot.delete_message(chat_id=chat_id, message_id=message_id)
             context.bot.send_message(chat_id=chat_id, text=text)
